Refactor : 코드리뷰 반영
코드리뷰를 반영하여 수정하였습니다.
